refactor: functions > components
